pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cryptkit"
version = "0.1.0"
description = "A smart CTF crypto & steganography assistant"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "Pillow>=10.0",
    "piexif>=1.1",
]

[project.scripts]
cryptkit = "cli:main"

[tool.setuptools]
py-modules = ["cli"]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "solvers*", "utils*"]
